[FLINK-16742][runtime] Refactor BashJavaUtils, move to separated package and extract configuration loading logics.
