+++ flink-runtime/src/main/java/org/apache/flink/runtime/util/bash/BashJavaUtils.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.util;
+package org.apache.flink.runtime.util.bash;
 
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.configuration.Configuration;
@@ -24,18 +24,11 @@ import org.apache.flink.configuration.JobManagerOptions;
 import org.apache.flink.configuration.TaskManagerOptions;
 import org.apache.flink.runtime.clusterframework.TaskExecutorProcessSpec;
 import org.apache.flink.runtime.clusterframework.TaskExecutorProcessUtils;
-import org.apache.flink.runtime.entrypoint.ClusterConfigurationParserFactory;
 import org.apache.flink.runtime.jobmanager.JobManagerProcessSpec;
 import org.apache.flink.runtime.jobmanager.JobManagerProcessUtils;
 import org.apache.flink.runtime.util.config.memory.ProcessMemoryUtils;
-import org.apache.flink.util.FlinkException;
 
-import org.apache.commons.cli.Options;
-
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Iterator;
-import java.util.List;
 
 import static org.apache.flink.util.Preconditions.checkArgument;
 
@@ -77,7 +70,7 @@ public class BashJavaUtils {
 	}
 
 	private static Configuration getConfigurationForStandaloneTaskManagers(String[] args) throws Exception {
-		Configuration configuration = loadConfiguration(args);
+		Configuration configuration = FlinkConfigLoader.loadConfiguration(args);
 		return TaskExecutorProcessUtils.getConfigurationMapLegacyTaskManagerHeapSizeToConfigOption(
 			configuration, TaskManagerOptions.TOTAL_FLINK_MEMORY);
 	}
@@ -89,37 +82,9 @@ public class BashJavaUtils {
 	}
 
 	private static Configuration getConfigurationForStandaloneJobManager(String[] args) throws Exception {
-		Configuration configuration = loadConfiguration(args);
+		Configuration configuration = FlinkConfigLoader.loadConfiguration(args);
 		return JobManagerProcessUtils.getConfigurationWithLegacyHeapSizeMappedToNewConfigOption(
 			configuration, JobManagerOptions.TOTAL_FLINK_MEMORY);
-	}
-
-	@VisibleForTesting
-	static Configuration loadConfiguration(String[] args) throws FlinkException {
-		return ConfigurationParserUtils.loadCommonConfiguration(
-			filterCmdArgs(args),
-			BashJavaUtils.class.getSimpleName());
-	}
-
-	private static String[] filterCmdArgs(String[] args) {
-		final Options options = ClusterConfigurationParserFactory.options();
-		final List<String> filteredArgs = new ArrayList<>();
-		final Iterator<String> iter = Arrays.asList(args).iterator();
-
-		while (iter.hasNext()) {
-			String token = iter.next();
-			if (options.hasOption(token)) {
-				filteredArgs.add(token);
-				if (options.getOption(token).hasArg() && iter.hasNext()) {
-					filteredArgs.add(iter.next());
-				}
-			} else if (token.startsWith("-D")) {
-				// "-Dkey=value"
-				filteredArgs.add(token);
-			}
-		}
-
-		return filteredArgs.toArray(new String[0]);
 	}
 
 	/**
